pyproject.toml
[tool.poetry]
name = "nonebot-plugin-guess"
version = "0.1.0"
description = "nobebot2插件《猜猜看》"
authors = ["freemt"]
repository = "https://github.com/ffreemt/nonebot-plugin-guess-game"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "2.0.0a10"
logzero = "^1.6.3"
pydantic = "1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.21.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
